# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "PIII- test",
    "description": "Aplicacion inicial de backend",
    "version": "1.0.0"
  },
  "host": "localhost:6000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/user": {
      "get": {
        "tags": [
          "Usuario"
        ],
        "description": "",
        "parameters": [
          {
            "name": "params",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Usuario"
        ],
        "description": "",
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/{id}": {
      "get": {
        "tags": [
          "Usuario"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Usuario"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Usuario"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/task": {
      "get": {
        "summary": "Obtener todas las tareas",
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Task"
              }
            }
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      },
      "post": {
        "summary": "Crear una nueva tarea",
        "parameters": [
          {
            "name": "task",
            "in": "body",
            "description": "Datos de la tarea",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Task"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Tarea creada exitosamente",
            "schema": {
              "$ref": "#/definitions/Task"
            }
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/task/{id}": {
      "get": {
        "summary": "Obtener una tarea por ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID de la tarea",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "schema": {
              "$ref": "#/definitions/Task"
            }
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      },
      "put": {
        "summary": "Actualizar una tarea",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID de la tarea",
            "required": true,
            "type": "string"
          },
          {
            "name": "task",
            "in": "body",
            "description": "Datos actualizados de la tarea",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Task"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tarea actualizada exitosamente",
            "schema": {
              "$ref": "#/definitions/Task"
            }
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      },
      "delete": {
        "summary": "Eliminar una tarea",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID de la tarea",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Tarea eliminada exitosamente"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    }
  },
  "definitions": {
    "Task": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "description": "ID único de la tarea"
        },
        "title": {
          "type": "string",
          "description": "Título de la tarea"
        },
        "description": {
          "type": "string",
          "description": "Descripción de la tarea"
        },
        "completed": {
          "type": "boolean",
          "description": "Indica si la tarea está completada o no"
        },
        "user": {
          "type": "string",
          "description": "ID del usuario asociado a la tarea"
        }
      }
    }
  }
}